+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
+## 1.2.0
+- Add option to specify the 2x suffix
+- Add option to bypass requirement of 2x suffix when using a transform
+
 ## 1.1.0
-- Use specified transform settings when using a transform.
+- Use specified transform settings when using a transform
 
 ## 1.0.1
 - Fix issue when using transforms *(rebased)*